refactor(apex-lsp-compliant-services): extract cross-file resolution substrate - W-24145209
Move the cross-file resolution layer into src/worker/crossFileResolution.ts:
recompileCursorFileAtFullDetail, loadReferencedTypesForFile,
resolveMissingNamesViaDataOwner and its private hasCompleteLocalSymbolTable.
The shared full-detail cursor cache they and the composition-root enrichment
path both touch moves to a small leaf, src/worker/fullDetailCursorCache.ts.

The module depends only on runtimeContext (assistance transport), the cache
leaf and external packages, so it is a clean mid-layer. The rename subsystem
now imports recompile / loadReferencedTypes from here.

With this, worker/rename.ts no longer imports the composition root at all:
the last back-edge is gone and the worker module graph is a proper acyclic DAG
(leaves -> cursorResolution / crossFileResolution -> rename -> composition
root). The shared module imports the substrate and re-exports the
previously-public surface unchanged.

W-24145209
